refactor(kimi-code): render the swarm dashboard in a dedicated SwarmCard
The swarm dashboard was stacked as branches inside ToolCallComponent,
against the guideline that new tool-result display should live behind a
dedicated renderer/component boundary. The static tool-renderer registry
can't host it (the dashboard needs live, per-event updates on a single
stable component — folding it in was how the original duplicate-card bug
was fixed), so extract it into a dedicated top-level SwarmCard selected
at tool-call-start, hosted by the same managed lifecycle.

- New SwarmCard (sibling to ToolCallComponent) + a narrow ManagedToolCard
  interface the streaming-UI registry is typed against; shared helpers
  (str/formatTokens/SWARM_ACTIVITY_MAX_LENGTH) moved to tool-call-shared.
- streaming-ui selects SwarmCard for name==='Swarm' at the single creation
  point; replay routes Swarm entries to SwarmCard too.
- session-event-handler narrows the generic subagent path to
  ToolCallComponent after the swarm guard.
- ToolCallComponent loses all swarm code (isSwarm() now returns false).

Behavior-preserving: same single stable card, in-place mutation, static
bullet, header from live args. Full suite green; swarm tests retargeted
to SwarmCard.
